Convert some more xla ops to be generated from a spec. (#1078)
Ops include: cat diagonal_value dynamic_slice dynamic_update_slice nll_loss
permute_value physical_cast prod repeat resize_value round_to_even select stack

+++ Sources/CX10/xla_tensor_wrapper.h
@@ -332,9 +332,9 @@ XLA_API OpaqueXLATensor* XLATensor_mul(OpaqueXLATensor* a, OpaqueXLATensor* b);
 XLA_API OpaqueXLATensor* XLATensor_mm(OpaqueXLATensor* a, OpaqueXLATensor* b);
 XLA_API OpaqueXLATensor* XLATensor_ne(OpaqueXLATensor* a, OpaqueXLATensor* b);
 XLA_API OpaqueXLATensor* XLATensor_neg(OpaqueXLATensor* a);
-XLA_API OpaqueXLATensor*
-XLATensor_nll_loss(OpaqueXLATensor* input, OpaqueXLATensor* target,
-                   int ignore_index);
+XLA_API OpaqueXLATensor* XLATensor_nll_loss(OpaqueXLATensor* input,
+                                            OpaqueXLATensor* target,
+                                            int64_t ignore_index);
 XLA_API OpaqueXLATensor*
 XLATensor_permute_value(OpaqueXLATensor* a, Int64ArrayRef arr);
 XLA_API OpaqueXLATensor* XLATensor_physical_cast(
